Take into account that chars are signed also in 32-bit x86
Previously, our code had different paths to deal with the fact that in
the C world, chars are signed in x86_64. However, chars are signed also
in x86 (i.e., the 32-bit platform), and this case we were not handling.

This commit simply makes sure we handle both x86 and x86_64 in the same
way: by using the `i8` type for chars.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch