Update dependencies to fix security vulnerabilities
The only required adjustment in the code itself was in src/init.rs. In
the older version of `nix`, `Result<T>` was defined as `Result<T,
Error>`, and `Error` had to be converted to an `Errno` with the `
as_errno()` method.

In the newer version, `Result<T>` is defined as `Result<T, Errno>`. We
get the `Errno` directly, so we don't need to call `as_errno()` anymore.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch