Allow running with unsupported target device types
This adds the `--no-dt-check` option, which skips the checks for
compatibility between the actual device and the device type of the
target fleet. It will also skip a similar check by device type when
downloading the balenaOS image.

This is useful for running takeover on devices that are technically
compatible with the fleet device type, but are not officially
tested/supported.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: minor