Give more time to processes to react to SIGTERM
In some test executions, we noticed warnings in the logs telling that
some processes (notably, the balenaEngine daemon) were still running
after the 500ms period we gave them to react to SIGTERM.

From our experience, 500ms seems indeed a bit too short for
balenaEngine, particularly when it's running a lot of containers that
need to be individually stopped before the daemon itself can exit.

We may want to try some smarter approach in the future, but for now, we
are simply increasing the timeout to 2 seconds. This shall be enough for
all processes to stop, and will not make any significant impact in the
overall time the whole takeover tool takes to run.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch