Make copy of `telinit` if a symlink to `init`
At the end of stage1, we bind-mount `takeover` on top of `init`, and
then call `telinit u`. Unfortunately, in some distributions (like
Devuan) `telinit` is a symlink to `init`. In this case, we lose access
to `telinit`, because the symlink now points to what is effectively
`takeover`.

With this commit, we check if `telinit` is a symlink to `init`, and if
so, we make a safe copy of `telinit` so that we can call it after the
bind-mounting takes place.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch