Log process list before killing processes
This is for better troubleshooting.

In one of my test runs, takeover failed because it was unable to kill
one of the processes during stage2. The logs contained only the PID of
the process, which was not enough to form any hypothesis about what
could have failed.

We already had code that would log a nice process table including the
program names, but strangely this was executed only when and if all
processes were killed successfully.

This commit:

1. Factor out the process table logging into a separate function.
2. Call this function before trying to kill the processes.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch