Don't run DT-specific code with --no-dt-check
When we first added the --no-dt-check, we naively implemented `Dummy` as
a `Device` pretending to support all device types (DTs).

The problem with this approach is that when passing the `--no-dt-check`,
we'd run all DT-specific code paths (because it was a `Device` that
supports all DTs, right?!).

This commit changes `Dummy` to be a `Device` that supports no DTs at
all, such that DT-specific code paths are run only when we are running
with a real DT that needs specific handling.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>