Handle race: read_link when iterating over /proc/PID/fd/FD
Iterating over /proc and doing some operation over its entries is
inherently racy, because entries on the /proc filesystem can vanish at
any time as processes die.

In order to take this into account, this commit changes the error
handling of a `read_link()` call we make while iterating over the
processes. If the error is a "file not found", we now simply ignore it.
This is safe to do, because we are only skipping code that would try to
kill the process (which doesn't exist anymore anyway).

Other errors may represent something serious, so they are reported and
cause the migration to be aborted.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch