Handle race: reading /proc/PID/root
Short story: the value read from `/proc/PID/root` wasn't being used
anywhere, so this commit just got rid of this reading, and of the value
itself.

Longer story: the initial motivation for this change was a race
condition when doing this reading. Iterating over /proc and doing some
operation over its entries is inherently racy, because entries on the
/proc filesystem can vanish at any time as processes die. Therefore, the
reading would sometimes fail with "file not found", which would
needlessly and unfortunately abort the migration process.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch