Handle race: when reading /proc/PID/status
Iterating over /proc and doing some operation over its entries is
inherently racy, because entries on the /proc filesystem can vanish at
any time as processes die.

In order to take this into account, this commit changes the error
handling of a `read_to_string()` call we make while iterating over
`/proc` to print a list of processes.

We now simply ignore "file not found" errors, which is fine because the
status information about a process that has already exited will not be
used in any meaningful way.

Other errors may represent something serious, so they are reported and
cause the migration to be aborted.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch