Allow logging to internal disk as a workaround
The recommended way to persistent logging during the takeover process is
to provide an external disk or internal disk that's not the target for
balenaOS. This PR adds a fallback mechanism that keeps the log in RAM
and then dumps the logs to the disk running the os. This can be helpful
when running a takeover remotely. This mechanism is being added to allow
running as part of a balenaOS to balenaOS update whereby the disk
partition layout changes across OS versions updates.

*Note:* the current mechanism is only tested from balenaOS to balenaOS
migration.

- Adds a new CLI option `--fallback_log`. This is not compatible with
  other logging options takeover provides
- Adds a new CLI option `--fallback_log_filename`. Allows overriding the
  fallback log file name. Defaults to `fallback.log`
- Adds a new CLI option `--fallback_log_dir`. Allows overriding the
  directory name where fallback log is persisted. Defaults to
  `fallback_log`
- Dumps partial logs captured if the takeover process fails at any point
  during the different stages
- Logs are persisted on the data partition. The logs are at
  `/mnt/data/balenahup`

We use mod_logger `set_log_dest` rather than `set_log_file` because
`set_log_file` truncates existing files and in our implementation, we
want to keep using the same file across the different stages of the
migration. `set_log_dest` flushes the
buffer(https://docs.rs/mod_logger/0.8.4/src/mod_logger/lib.rs.html#222)
before actually writing to the stream provided, so we manually get the
buffered memory logs at the start of stage2-init and stage2 worker
process and then attempt to write it to the log file once it is set up.

Change-type: minor
Signed-off-by: Rahul Thakoor <rahul@balena.io>