Allow logging to internal disk as a workaround
The recommended way to persistent logging during the takeover process is
to provide an external disk or internal disk that's not the target for
balenaOS. This PR adds a mechanism that keeps the log in RAM and then
dumps the logs to the disk running the os. This can be helpful when
running a takeover remotely.

- Adds a new CLI option `--log-to-balenaos`. This is not compatible with
  other logging options takeover provides
- Dumps partial logs at captured if the takeover process fails at any
  point during the different stages

This mechanism is being added to allow running as part of a balenaOS to
balenaOS update whereby the disk partition layout changes across OS versions.

Change-type: minor
Signed-off-by: Rahul Thakoor <rahul@balena.io>